Fixed exports.

+++ installer/installer.sh
@@ -54,18 +54,25 @@ e.g. nvme0n1p1
 
     # format and mount efi partition
     printf "\n\n"
-    lsblk -f
-    read -p "EFI PARTITION: " PARTITION
-    while true
-    do
-        [ -b "/dev/$PARTITION" ] || ( read -p "INVALID, TRY AGAIN. EFI PARTITION: " PARTITION; continue )
-        mkfs.fat -F 32 "/dev/$PARTITION"
-        mount --mkdir "/dev/$PARTITION" "/mnt/efi"
+	read -p "Do you want to use EFI partition? [y]es [n]o" ANSWER
 
-        [ "$?" = 0 ] && (printf "\nPartition mounted!\n\n") || \
-          (printf "Something went wrong!")
-        break;
-    done
+	if [ "$ANSWER" = 'y' ]
+		then
+		lsblk -f
+    		read -p "EFI PARTITION: " PARTITION
+    		while true
+    		do
+			[ ! -b "/dev/$PARTITION" ] && ( read -p "INVALID, TRY AGAIN. EFI PARTITION: " PARTITION; continue )
+			read -p "DO YOU WANT TO FORMAT EFI PARTITION? [y]es [n]o " ANSWER
+
+        		[ "$ANSWER" = "y" ] && mkfs.fat -F 32 "/dev/$PARTITION"
+        		mount --mkdir "/dev/$PARTITION" "/mnt/efi"
+
+		        [ "$?" = 0 ] && (printf "\nPartition mounted!\n\n") || 
+		        (printf "Something went wrong!")
+        		break;
+    		done
+	fi
 
 	# format and mount swap partition
 	printf "\n\n"
@@ -99,9 +106,9 @@ e.g. nvme0n1p1
 
 	# install system
 	printf "\n\nBeginning the system installation\n"
-	pacstrap /mnt base-devel linux systemd-sysvcompat iputils git grub efibootmgr ntfs-3g --noconfirm
+	pacstrap /mnt base-devel linux systemd-sysvcompat iputils networkmanager vim git grub efibootmgr ntfs-3g  --noconfirm
 
-	[ $? != 0 ] && echo "SORRY! THE INSTALLATION PROCESS FAILED!"
+	[ $? != 0 ] && ( echo "SORRY! THE INSTALLATION PROCESS FAILED! EXITING..."; return 1 )
 
 	printf "
 
@@ -110,11 +117,11 @@ THE INSTALLATION PROCESS HAS FINISHED SUCCESSFULLY!
 NOW LET'S CONFIGURE IT...
 
 "
-	cp ~/installers/configurator.sh /mnt/
-	chmod 0775 /mnt/configurator.sh
+	cp -r installer/ /mnt/
+	chmod 0775 /mnt/installer/configurator.sh
 
 	genfstab -L /mnt >> /mnt/etc/fstab
-	arch-chroot /mnt ./configurator.sh
+	arch-chroot /mnt ./installer/configurator.sh
 
 	if [ "$?" = 0 ]
 	then
